finally vanilla style
phew, I almost lost all my progress in this commit! Made a checkout from a previous commit and a detached head was created which I wasn't able to publish to the main branch. I then switched to the main branch and all my changes was gone. Thankfully I had ctrl-z.